requirements.txt
fastapi>= 0.110.0,<0.110.1
uvicorn>=0.28.0,<0.29.0
pyjwt